Rename telemetry modules
