+++ src/telemetry/hardware_info.rs
